Initial checkpoint of poudriere-as-a-service
The initial goal of paas is to spawn AWS instances, bootstrap a
poudriere configuration, build select packages, and make them
available somewhow (e.g., via EBS).  Configuration is via userdata
and prototype scripts to be included in the userdata tarball are
in the userdata directory.

An eventual goal is to be able to do things like orchestrate ports
exp-runs by running two simultaneous poudriere builds with different
configurations or to do a ports-universe run across multiple
architectures or OS configs.  Doing so efficently will require
considerable optimization.

The current commit is just a collection of poorly connected experiments
to see what can work. A proper design with a configuration language is
required.